Rename files: s/stick10/NKPro and s/stick20/NKStorage
