+++ include/NKPro_commands_0.8.h
@@ -12,7 +12,7 @@
 #include <stdint.h>
 #include "command.h"
 #include "device_proto.h"
-#include "stick10_commands.h"
+#include "NKPro_commands.h"
 
 namespace nitrokey {
     namespace proto {
